add political party happiness from social classes happiness

+++ packages/server/database/migrations/1736026116323_create_social_class_happiness_modifiers_table.ts
@@ -12,7 +12,7 @@ export default class extends BaseSchema {
       table.string('color').notNullable();
       table.enum('type', ['Permanent', 'Temporary']).notNullable();
       table.integer('duration').nullable();
-      table.integer('amount').notNullable();
+      table.float('amount').notNullable();
 
       table
         .integer('social_class_id')
